move destroy to sugar context
to be shared by both Cython and cffi

fixes bugs in Cython implementation
