+++ zmq/backend/cython/context.pyx
@@ -34,7 +34,6 @@ cdef class Context:
     
     def __cinit__(self, int io_threads=1, size_t shadow=0, **kwargs):
         self.handle = NULL
-        self._sockets = NULL
         if shadow:
             self.handle = <void *>shadow
             self._shadow = True
@@ -44,70 +43,26 @@ cdef class Context:
                 self.handle = zmq_ctx_new()
             else:
                 self.handle = zmq_init(io_threads)
-        
+
         if self.handle == NULL:
             raise ZMQError()
-        
+
         cdef int rc = 0
         if ZMQ_VERSION_MAJOR >= 3 and not self._shadow:
             rc = zmq_ctx_set(self.handle, ZMQ_IO_THREADS, io_threads)
             _check_rc(rc)
-        
+
         self.closed = False
-        self._n_sockets = 0
-        self._max_sockets = 32
-        
-        self._sockets = <void **>malloc(self._max_sockets*sizeof(void *))
-        if self._sockets == NULL:
-            raise MemoryError("Could not allocate _sockets array")
-        
         self._pid = getpid()
-    
+
     def __dealloc__(self):
         """don't touch members in dealloc, just cleanup allocations"""
         cdef int rc
-        if self._sockets != NULL:
-            free(self._sockets)
-            self._sockets = NULL
-            self._n_sockets = 0
 
         # we can't call object methods in dealloc as it
         # might already be partially deleted
         if not self._shadow:
             self._term()
-    
-    cdef inline void _add_socket(self, void* handle):
-        """Add a socket handle to be closed when Context terminates.
-        
-        This is to be called in the Socket constructor.
-        """
-        if self._n_sockets >= self._max_sockets:
-            self._max_sockets *= 2
-            self._sockets = <void **>realloc(self._sockets, self._max_sockets*sizeof(void *))
-            if self._sockets == NULL:
-                raise MemoryError("Could not reallocate _sockets array")
-        
-        self._sockets[self._n_sockets] = handle
-        self._n_sockets += 1
-
-    cdef inline void _remove_socket(self, void* handle):
-        """Remove a socket from the collected handles.
-        
-        This should be called by Socket.close, to prevent trying to
-        close a socket a second time.
-        """
-        cdef bint found = False
-        
-        for idx in range(self._n_sockets):
-            if self._sockets[idx] == handle:
-                found=True
-                break
-        
-        if found:
-            self._n_sockets -= 1
-            if self._n_sockets:
-                # move last handle to closed socket's index
-                self._sockets[idx] = self._sockets[self._n_sockets]
 
     @property
     def underlying(self):
@@ -212,35 +167,5 @@ cdef class Context:
 
         return rc
 
-    def destroy(self, linger=None):
-        """ctx.destroy(linger=None)
-        
-        Close all sockets associated with this context, and then terminate
-        the context. If linger is specified,
-        the LINGER sockopt of the sockets will be set prior to closing.
-        
-        .. warning::
-        
-            destroy involves calling ``zmq_close()``, which is **NOT** threadsafe.
-            If there are active sockets in other threads, this must not be called.
-        """
-        
-        cdef int linger_c
-        cdef bint setlinger=False
-        
-        if linger is not None:
-            linger_c = linger
-            setlinger=True
 
-        if self.handle != NULL and not self.closed and self._n_sockets:
-            while self._n_sockets:
-                if setlinger:
-                    zmq_setsockopt(self._sockets[0], ZMQ_LINGER, &linger_c, sizeof(int))
-                rc = zmq_close(self._sockets[0])
-                if rc < 0 and zmq_errno() != ZMQ_ENOTSOCK:
-                    raise ZMQError()
-                self._n_sockets -= 1
-                self._sockets[0] = self._sockets[self._n_sockets]
-        self.term()
-    
 __all__ = ['Context']
